+++ src/index.ts
@@ -1,3 +1,3 @@
 export {default as Gizmo} from './Gizmo/Gizmo'
 export {defaultGizmoOptions} from './Gizmo/constants'
-export {GizmoOptions} from './Gizmo/types'
+export {type GizmoOptions} from './Gizmo/types'